Don't drop executables with password protected zip sections

+++ extract/extract.py
@@ -128,7 +128,9 @@ class Extract(ServiceBase):
         if num_extracted == 0 and password_protected:
             section = ResultSection("Failed to extract password protected file.", heuristic=Heuristic(12))
             section.add_tag('file.behavior', "Archive Unknown Password")
-            request.drop()
+            if not request.file_type.startswith("executable"):
+                # Don't drop executables that contain password protected zip sections
+                request.drop()
 
         elif num_extracted != 0:
             if password_protected and self._last_password is not None:
